Re-use types
